requirements.txt
requests~=2.31.0
pillow~=10.2.0
numpy~=1.26.4
pytube~=15.0.0
setuptools~=65.5.0